subscriber-service/requirements.txt
flask
confluent-kafka
python-dotenv
requests